Add files via upload
This project we created a nested for loop obtain a student's name and five grades from each of them for two arrays to catch each of them (Student (String) & Grades (Double)). Soon after two different nested for loop were made to get the sum and average of the five grades. Then finally another nested for loop was created display the student's name, their five grades, and the average of the five grades.